[consensus] move consensus_types into its own crate
The plan is to move safety_rules into its own crate / binary, so that
we can minimize dependencies and have a clear crate-based dependency
graph. The first step to doing that is identifying cyclic dependencies
between safety_rules and consensus. The most obvious is consensus_types
/ common. This pr moves these files into a new crate.

Along the way there were a few compromises:
- Needed a wrapper around BlockSchema::Block as we cannot implement a
trait for an external generic
- Moved common into consensus_types since it fits the bill
- Cleaned up the cargo file a little bit
- Reimported the new crate into consensus using rust magic so that each
use of the types doesn't need to know it is now an external crate (this
should be fixed in a future PR)
