+++ Cargo.toml
@@ -22,6 +22,7 @@ members = [
     "config/config-builder",
     "config/generate-keypair",
     "consensus",
+    "consensus/consensus_types",
     "crypto/crypto",
     "crypto/crypto_derive",
     "crypto/secret-service",
