+++ consensus/consensus_types/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "consensus-types"
+version = "0.1.0"
+authors = ["Libra Association <opensource@libra.org>"]
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+bytes = "0.4.12"
+mirai-annotations = { version = "1.4.0", default-features = false }
+proptest = { version = "0.9.4", optional = true }
+protobuf = "~2.7"
+rand = { version = "0.6.5", default-features = false }
+rmp-serde = { version = "0.13.7", default-features = false }
+serde = { version = "1.0.99", default-features = false }
+
+canonical_serialization = { path = "../../common/canonical_serialization" }
+config = { path = "../../config" }
+crypto = { path = "../../crypto/crypto" }
+executor = { path = "../../execution/executor" }
+failure = { path = "../../common/failure_ext", package = "failure_ext" }
+network = { path = "../../network" }
+libra-types = { path = "../../types" }
+
+[dev-dependencies]
+proptest = "0.9.4"
+
+crypto = { path = "../../crypto/crypto", features = ["testing"]}
+libra-types = { path = "../../types", features = ["testing"]}
+
+[features]
+default = []
+fuzzing = []
+testing = ["proptest"]
