+++ consensus/consensus_types/src/lib.rs
@@ -1,0 +1,17 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+#[macro_use]
+extern crate failure;
+
+pub mod block;
+pub mod common;
+pub mod proposal_msg;
+pub mod quorum_cert;
+pub mod sync_info;
+#[cfg(any(test, feature = "testing"))]
+pub mod test_utils;
+pub mod timeout_certificate;
+pub mod timeout_msg;
+pub mod vote_data;
+pub mod vote_msg;
