--- consensus/src/chained_bft/consensus_types/mod.rs
@@ -1,11 +1,0 @@
-// Copyright (c) The Libra Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-pub(crate) mod block;
-pub(crate) mod proposal_msg;
-pub(crate) mod quorum_cert;
-pub(crate) mod sync_info;
-pub(crate) mod timeout_certificate;
-pub(crate) mod timeout_msg;
-pub(crate) mod vote_data;
-pub(crate) mod vote_msg;
