+++ consensus/src/chained_bft/consensusdb/schema/block/mod.rs
@@ -17,7 +17,7 @@ use prost::Message;
 use prost_ext::MessageExt;
 use schemadb::schema::{KeyCodec, Schema, ValueCodec};
 use std::convert::TryInto;
-use std::marker::PhantomData;
+use std::{cmp, fmt, marker::PhantomData};
 
 pub struct BlockSchema<T: Payload> {
     phantom: PhantomData<T>,
@@ -26,7 +26,7 @@ pub struct BlockSchema<T: Payload> {
 impl<T: Payload> Schema for BlockSchema<T> {
     const COLUMN_FAMILY_NAME: schemadb::ColumnFamilyName = BLOCK_CF_NAME;
     type Key = HashValue;
-    type Value = Block<T>;
+    type Value = SchemaBlock<T>;
 }
 
 impl<T: Payload> KeyCodec<BlockSchema<T>> for HashValue {
@@ -39,14 +39,43 @@ impl<T: Payload> KeyCodec<BlockSchema<T>> for HashValue {
     }
 }
 
-impl<T: Payload> ValueCodec<BlockSchema<T>> for Block<T> {
+#[derive(Clone)]
+/// SchemaBlock is a crate wrapper for Block that is defined outside this crate.
+/// ValueCodec cannot be implemented for Block here as Block is defined in
+/// consensus_types crate (E0210).
+pub struct SchemaBlock<T: Payload>(Block<T>);
+
+impl<T: Payload> SchemaBlock<T> {
+    pub fn from_block(sb: Block<T>) -> SchemaBlock<T> {
+        Self(sb)
+    }
+
+    pub fn borrow_into_block(&self) -> &Block<T> {
+        &self.0
+    }
+}
+
+impl<T: Payload> cmp::PartialEq for SchemaBlock<T> {
+    fn eq(&self, other: &SchemaBlock<T>) -> bool {
+        self.borrow_into_block().eq(&other.borrow_into_block())
+    }
+}
+
+impl<T: Payload> fmt::Debug for SchemaBlock<T> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.borrow_into_block().fmt(f)
+    }
+}
+
+impl<T: Payload> ValueCodec<BlockSchema<T>> for SchemaBlock<T> {
     fn encode_value(&self) -> Result<Vec<u8>> {
-        let block: network::proto::Block = self.clone().into();
+        let block: network::proto::Block = self.borrow_into_block().clone().into();
         Ok(block.to_vec()?)
     }
 
     fn decode_value(data: &[u8]) -> Result<Self> {
-        network::proto::Block::decode(data)?.try_into()
+        let block: Block<T> = network::proto::Block::decode(data)?.try_into()?;
+        Ok(Self::from_block(block))
     }
 }
